moved getRootIndex
